requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.6.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Authentication (using PyJWT directly)
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Environment
python-dotenv>=1.0.0